feat(21): refactor Dashboard to use custom useDashboard hook for event management
