+++ app/dashboard/page.tsx
@@ -1,74 +1,10 @@
 "use client";
 
 import { Calendar } from "@/components/ui/calendar";
-import { CalendarEvent } from "@/types/event";
-import React, { useMemo } from "react";
-import { GETEvents } from "@/network/google_calendar";
-import { createClient } from "@/utils/supabase/client";
-import { mapGoogleCalendarEvents } from "@/utils/mappers/mapGoogleCalendarEvents";
-import type { GoogleCalendarEvent } from "@/utils/mappers/mapGoogleCalendarEvents";
+import { useDashboard } from "@/hooks/dashboard/useDashboard";
 
 export default function Dashboard() {
-  const [date, setDate] = React.useState<Date | undefined>(new Date());
-  const [events, setEvents] = React.useState<CalendarEvent[]>([]);
-  const [loading, setLoading] = React.useState<boolean>(true);
-
-  const startOfMonth = new Date();
-  startOfMonth.setDate(1);
-  startOfMonth.setHours(0, 0, 0, 0);
-
-  const endOfMonth = new Date(startOfMonth);
-  endOfMonth.setMonth(endOfMonth.getMonth() + 1);
-
-  const getUserToken = async () => {
-    const supabase = createClient();
-    const {
-      data: { session },
-    } = await supabase.auth.getSession();
-    const token = session?.provider_token;
-    return token;
-  };
-
-  const fetchEvents = React.useCallback(async () => {
-    const params = new URLSearchParams({
-      timeMin: startOfMonth.toISOString(),
-      timeMax: endOfMonth.toISOString(),
-      singleEvents: "true",
-      orderBy: "startTime",
-    });
-
-    try {
-      const token = await getUserToken();
-      if (token) {
-        const calendarEvents = await GETEvents(token, params).then((response) =>
-          response.json()
-        ).then(({ items }: { items: GoogleCalendarEvent[]}): CalendarEvent[] =>{
-
-          const data = mapGoogleCalendarEvents(items);
-          return data;
-        });
-          setEvents(calendarEvents || []);
-      }
-    
-    } catch (error) {
-      console.error("Errore nel recupero degli eventi del calendario:", error);
-    } finally {
-      setLoading(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  React.useEffect(() => {
-    fetchEvents();
-  }, [fetchEvents]);
-
-  const eventsForSelectedDate = useMemo(() => {
-    return events.filter((event) => {
-      if (!date || !event.start?.dateTime) return false;
-      const eventDate = new Date(event.start.dateTime);
-      return eventDate.toDateString() === date.toDateString();
-    });
-  }, [events, date]);
+  const { date, eventsForSelectedDate, setDate, loading, hasEvents } = useDashboard()
 
   if (loading) {
     return <div className="p-6">Caricamento eventi del calendario...</div>;
@@ -84,13 +20,7 @@ export default function Dashboard() {
             selected={date}
             onSelect={setDate}
             modifiers={{
-              hasEvents: (day) => {
-                return events.some((event) => {
-                  if (!event.start?.dateTime) return false;
-                  const eventDate = new Date(event.start.dateTime);
-                  return eventDate.toDateString() === day.toDateString();
-                });
-              },
+              hasEvents
             }}
             modifiersClassNames={{
               hasEvents:
